+++ app/javascript/vue/tasks/matrix_image/matrix_image/components/View/Main.vue
@@ -5,7 +5,7 @@
       <thead>
         <tr>
           <th class="object-cell"/>
-          <th 
+          <th
             v-for="descriptor in descriptors"
             class="header-cell"
             :key="descriptor.id">
@@ -19,9 +19,11 @@
         <tr
           class="row-cell"
           v-for="(row, index) in rows">
-          <td
-            class="object-cell"
-            v-html="row.object.object_tag"/>
+          <td class="object-cell">
+            <a
+              v-html="row.object.object_tag"
+              :href="browseOtu(row.object.id)"/>
+          </td>
           <td
             v-for="rCol in row.depictions"
             class="padding-cell">
@@ -41,6 +43,8 @@
 
 import ajaxCall from 'helpers/ajaxCall'
 import SpinnerComponent from 'components/spinner'
+import { RouteNames } from 'routes/routes'
+import depiction from 'components/radials/object/images/depiction'
 
 export default {
   components: {
@@ -49,9 +53,14 @@ export default {
   props: {
     matrixId: {
       type: [Number, String],
-      required: true
+      default: undefined
+    },
+    otusId: {
+      type: [String, Array],
+      default: () => []
     }
   },
+
   data () {
     return {
       descriptors: [],
@@ -60,19 +69,30 @@ export default {
       isLoading: false
     }
   },
+
   created () {
     this.isLoading = true
-    ajaxCall('get', `/tasks/observation_matrices/image_matrix/${this.matrixId}/key`).then(({ body }) => {
+    const retrieveDepictions = this.otusId.length
+      ? ajaxCall('get', '/tasks/observation_matrices/image_matrix/0/key', { params: { otu_filter: this.otusId } })
+      : ajaxCall('get', `/tasks/observation_matrices/image_matrix/${this.matrixId}/key`)
+
+    retrieveDepictions.then(({ body }) => {
       this.descriptors = Object.values(body.list_of_descriptors)
-      this.rows = Object.values(body.depiction_matrix).filter(row => [].concat(...row.depictions).length)
-      this.isLoading = false
-    }).then(() => {
+      this.rows = Object.values(body.depiction_matrix)
+        .filter(row => [].concat(...row.depictions).length)
+        .map(observation => ({
+          ...observation,
+          depictions: observation.depictions.map(obsDepictions => obsDepictions.filter(depiction => depiction.depiction_object_type === 'Observation'))
+        }))
+    }).finally(() => {
       this.isLoading = false
     })
+  },
+
+  methods: {
+    browseOtu (id) {
+      return `${RouteNames.BrowseOtu}?otu_id=${id}`
+    }
   }
 }
 </script>
-
-<style>
-
-</style>